+++ src/src/PluginStructs/P128_data_struct.cpp
@@ -868,7 +868,7 @@ bool P128_data_struct::plugin_fifty_per_second(struct EventStruct *event) {
 
 void P128_data_struct::fade(void) {
   for (int pixel = 0; pixel < pixelCount; pixel++) {
-    long  counter  = 20 * (counter20ms - starttime[pixel]);
+    int32_t  counter  = 20 * (counter20ms - starttime[pixel]);
     float progress = (float)counter / (float)fadetime;
     progress = constrain(progress, 0.0f, 1.0f);
 
@@ -1033,7 +1033,7 @@ void P128_data_struct::faketv(void) {
  * Cycles a rainbow over the entire string of LEDs.
  */
 void P128_data_struct::rainbow(void) {
-  const long  counter  = 20 * (counter20ms - starttimerb);
+  const int32_t  counter  = 20 * (counter20ms - starttimerb);
   const float progress = (float)counter / (float)fadetime;
 
   if (fadeIn == true) {
